# TT-12965: Improve performance of ctx.GetOASDefinition/GetDefinition