--- a/apidef/oas/oas.go
+++ b/apidef/oas/oas.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/TykTechnologies/tyk/apidef"
 	"github.com/TykTechnologies/tyk/config"
+	"github.com/TykTechnologies/tyk/internal/reflect"
 )
 
 const (
@@ -200,6 +201,11 @@ func (s *OAS) RemoveTykExtension() {
 	}
 
 	delete(s.Extensions, ExtensionTykAPIGateway)
+}
+
+// Clone creates a deep copy of the OAS object and returns a new instance.
+func (s *OAS) Clone() (*OAS, error) {
+	return reflect.Clone(s), nil
 }
 
 func (s *OAS) getTykAuthentication() (authentication *Authentication) {