--- a/ctx/ctx.go
+++ b/ctx/ctx.go
@@ -5,16 +5,15 @@ import (
 	"encoding/json"
 	"net/http"
 
-	"github.com/TykTechnologies/tyk/internal/httputil"
-
-	"github.com/TykTechnologies/tyk/apidef/oas"
-
-	"github.com/TykTechnologies/tyk/config"
-
 	"github.com/TykTechnologies/tyk/apidef"
-	logger "github.com/TykTechnologies/tyk/log"
+	"github.com/TykTechnologies/tyk/apidef/oas"
+	"github.com/TykTechnologies/tyk/config"
+	"github.com/TykTechnologies/tyk/internal/httputil"
+	"github.com/TykTechnologies/tyk/internal/reflect"
 	"github.com/TykTechnologies/tyk/storage"
 	"github.com/TykTechnologies/tyk/user"
+
+	logger "github.com/TykTechnologies/tyk/log"
 )
 
 type Key uint
@@ -57,7 +56,6 @@ const (
 )
 
 func ctxSetSession(r *http.Request, s *user.SessionState, scheduleUpdate bool, hashKey bool) {
-
 	if s == nil {
 		panic("setting a nil context SessionData")
 	}
@@ -84,7 +82,10 @@ func ctxSetSession(r *http.Request, s *user.SessionState, scheduleUpdate bool, h
 
 func GetAuthToken(r *http.Request) string {
 	if v := r.Context().Value(AuthToken); v != nil {
-		return v.(string)
+		value, ok := v.(string)
+		if ok {
+			return value
+		}
 	}
 	return ""
 }
@@ -114,45 +115,32 @@ func SetSession(r *http.Request, s *user.SessionState, scheduleUpdate bool, hash
 	}
 }
 
+// SetDefinition can be used to set a modified API definition to the request context.
 func SetDefinition(r *http.Request, s *apidef.APIDefinition) {
 	ctx := r.Context()
 	ctx = context.WithValue(ctx, Definition, s)
 	httputil.SetContext(r, ctx)
 }
 
+// GetDefinition will return a copy of the API definition valid for the request.
+// This copy is there to prevent shared API definition pointers.
 func GetDefinition(r *http.Request) *apidef.APIDefinition {
 	if v := r.Context().Value(Definition); v != nil {
 		if val, ok := v.(*apidef.APIDefinition); ok {
-			return val
-		} else {
-			logger.Get().Warning("APIDefinition struct differ from the gateway version, trying to unmarshal.")
-			def := apidef.APIDefinition{}
-			b, _ := json.Marshal(v)
-			e := json.Unmarshal(b, &def)
-			if e == nil {
-				return &def
-			}
+			return reflect.Clone(val)
 		}
 	}
+
 	return nil
 }
 
 // GetOASDefinition returns a deep copy of the OAS definition of the called API.
 func GetOASDefinition(r *http.Request) *oas.OAS {
-	v := r.Context().Value(OASDefinition)
-	if v == nil {
-		return nil
+	if v := r.Context().Value(OASDefinition); v != nil {
+		if val, ok := v.(*oas.OAS); ok {
+			return reflect.Clone(val)
+		}
 	}
 
-	val, ok := v.(*oas.OAS)
-	if !ok {
-		return nil
-	}
-
-	ret, err := val.Clone()
-	if err != nil {
-		logger.Get().WithError(err).Error("Cloning OAS object in the request context")
-	}
-
-	return ret
+	return nil
 }