The *oas.OAS or *apidef.APIDefinition retrieved directly from context should be considered immutable, and depending on how plugins use the ctx package, this may cause issues. Each server request is started in a goroutine, making any map iteration dangerous due to the concurrency. By providing a deep copy that problem goes away.

The PR optimizes performance for the following functions:

- ctx.GetDefinition

This function returned the pointer to APIDefinition directly as if you would have used the context. This was a known defect until now. We now use the same deep copy mechanism that was optimized for OAS, to avoid concurrency concerns from plugins and improve stability.

- ctx.GetOASDefinition

Since we were aware of the defect of GetDefinition, this function used a naive copy mechanism to ensure that the copy is safe to use. It did rely on encoding/json, on the reasoning performance can be improved by a more optimized mechanism in the future (encoding/gob, msgpack...). Using a memory cloning approach, the following change was benchmarked, showing a reduction of latencies and allocations around -90%.

Benchmark master:

```
BenchmarkOAS_Clone     	   47610	     25781 ns/op	   15211 B/op	     206 allocs/op
BenchmarkOAS_Clone-2   	   44457	     25821 ns/op	   15214 B/op	     206 allocs/op
BenchmarkOAS_Clone-4   	   51541	     24412 ns/op	   15220 B/op	     206 allocs/op
BenchmarkOAS_Clone-8   	   47432	     27178 ns/op	   15228 B/op	     206 allocs/op
```

Benchmark PR:

```
BenchmarkOAS_Clone     	  390181	      2839 ns/op	    1408 B/op	      17 allocs/op
BenchmarkOAS_Clone-2   	  455761	      2547 ns/op	    1408 B/op	      17 allocs/op
BenchmarkOAS_Clone-4   	  436094	      2570 ns/op	    1408 B/op	      17 allocs/op
BenchmarkOAS_Clone-8   	  447078	      3100 ns/op	    1408 B/op	      17 allocs/op
```

- Improves ctx.GetOASDefinition
- Improves ctx.GetDefinition to return a copy (concurrency fix)
- Additional benchmarks added for ctx package
- Additional benchmarks added for reflect package
- Added huandu/go-clone and an internal type-generic reflect.Clone API.

Closes #6471

<details open>
  <summary><a href="https://tyktech.atlassian.net/browse/TT-12965" title="TT-12965" target="_blank">TT-12965</a></summary>
  <br />
  <table>
    <tr>
      <th>Summary</th>
      <td>[Customer PR] Improve performance of ctx.GetOASDefinition()</td>
    </tr>
    <tr>
      <th>Type</th>
      <td>
        <img alt="Story" src="https://tyktech.atlassian.net/rest/api/2/universal_avatar/view/type/issuetype/avatar/10315?size=medium" />
        Story
      </td>
    </tr>
    <tr>
      <th>Status</th>
      <td>In Dev</td>
    </tr>
    <tr>
      <th>Points</th>
      <td>N/A</td>
    </tr>
    <tr>
      <th>Labels</th>
      <td><a href="https://tyktech.atlassian.net/issues?jql=project%20%3D%20TT%20AND%20labels%20%3D%20customer_bug%20ORDER%20BY%20created%20DESC" title="customer_bug">customer_bug</a>, <a href="https://tyktech.atlassian.net/issues?jql=project%20%3D%20TT%20AND%20labels%20%3D%20github%20ORDER%20BY%20created%20DESC" title="github">github</a>, <a href="https://tyktech.atlassian.net/issues?jql=project%20%3D%20TT%20AND%20labels%20%3D%20jira_escalated%20ORDER%20BY%20created%20DESC" title="jira_escalated">jira_escalated</a></td>
    </tr>
  </table>
</details>
<!--
  do not remove this marker as it will break jira-lint's functionality.
  added_by_jira_lint
-->

---

https://tyktech.atlassian.net/browse/TT-12965