@buger 2025-02-03
<p>:broken_heart: The detected issue is not in one of the allowed statuses :broken_heart: </p>    
   <table>
     <tr>
        <th>Detected Status</th>
        <td>Open</td>
        <td>:x:</td>
     </tr>
     <tr>
        <th>Allowed Statuses</th>
        <td>In Dev,In Code Review,Ready for Testing,In Test,In Progress,In Review</td>
        <td>:heavy_check_mark:</td>
      </tr>
   </table>
   <p>Please ensure your jira story is in one of the allowed statuses</p>