@buger 2025-02-11
<p>
    Let's make that PR title a 💯 shall we? 💪
    </p>
    <p>
    Your <em>PR title</em> and <em>story title</em> look <strong>slightly different</strong>. Just checking in to know if it was intentional!
    </p>
    <table>
      <tr>
        <th>Story Title</th>
        <td>[Customer PR] Improve performance of ctx.GetOASDefinition()</td>
      </tr>
      <tr>
          <th>PR Title</th>
          <td>TT-12965: Fix/improve ctx allocation with clone</td>
        </tr>
    </table>
    <p>
      Check out this <a href="https://www.atlassian.com/blog/git/written-unwritten-guide-pull-requests">guide</a> to learn more about PR best-practices.
    </p>